ingestion/requirements.txt
python-dotenv>=1.0.1
requests>=2.32.0
pandas>=2.2.2
pyarrow>=16.1.0
meteostat>=1.6.7
holidays>=0.52
pvlib>=0.10.5
tenacity>=8.3.0
fastparquet>=2024.2.0
beautifulsoup4>=4.12.3
lxml>=5.2.2
ratelimit>=2.2.1
tqdm>=4.66.4

